requirements.txt
line-bot-sdk
bs4
flask
matplotlib
datetime
imgurpython
urllib3==1.24.2
jieba==0.42.1
redis==3.4.1